+++ src/app/app.hpp
@@ -1,0 +1,14 @@
+#pragma once
+
+
+namespace shar {
+
+class App {
+public:
+  App();
+  ~App();
+
+  int run(int argc, char* argv[]);
+};
+
+}
