--- src/ui/app.hpp
@@ -1,14 +1,0 @@
-#pragma once
-
-
-namespace shar::ui {
-
-// main application state
-class App {
-public:
-
-private:
-
-};
-
-}
